create: activity export

+++ app/Http/Controllers/GapDisnakerController.php
@@ -81,17 +81,28 @@ class GapDisnakerController extends Controller
             $disnaker = GapDisnaker::find($id);
             $branch = Branch::find($request->branch_id);
             $jenis_perizinan = JenisPerizinan::find($request->jenis_perizinan_id);
-            $fileName = $request->file('file')->getClientOriginalName();
-            $request->file('file')->storeAs('infra/disnaker/'.$disnaker->id.'/', $fileName, ["disk" => 'public']);
-            Storage::disk('public')->delete('infra/disnaker/'.$disnaker->id.'/'.$disnaker->file);
-            $disnaker->update([
-                'branch_id' => $branch->id,
-                'jenis_perizinan_id' => $jenis_perizinan->id,
-                'tgl_pengesahan' => $request->tgl_pengesahan,
-                'tgl_masa_berlaku' => $request->tgl_masa_berlaku,
-                'progress_resertifikasi' => $request->progress_resertifikasi,
-                'file' => $fileName,
-            ]);
+            if (!is_null($request->file('file'))) {
+                $fileName = $request->file('file')->getClientOriginalName();
+                $request->file('file')->storeAs('infra/disnaker/' . $disnaker->id . '/', $fileName, ["disk" => 'public']);
+                Storage::disk('public')->delete('infra/disnaker/' . $disnaker->id . '/' . $disnaker->file);
+
+                $disnaker->update([
+                    'branch_id' => $branch->id,
+                    'jenis_perizinan_id' => $jenis_perizinan->id,
+                    'tgl_pengesahan' => $request->tgl_pengesahan,
+                    'tgl_masa_berlaku' => $request->tgl_masa_berlaku,
+                    'progress_resertifikasi' => $request->progress_resertifikasi,
+                    'file' => $fileName,
+                ]);
+            } else {
+                $disnaker->update([
+                    'branch_id' => $branch->id,
+                    'jenis_perizinan_id' => $jenis_perizinan->id,
+                    'tgl_pengesahan' => $request->tgl_pengesahan,
+                    'tgl_masa_berlaku' => $request->tgl_masa_berlaku,
+                    'progress_resertifikasi' => $request->progress_resertifikasi,
+                ]);
+            }
             return redirect(route('infra.disnaker'))->with(['status' => 'success', 'message' => 'Data Berhasil diupdate']);
         } catch (Throwable $e) {
             return redirect(route('infra.disnaker'))->with(['status' => 'failed', 'message' => $e->getMessage()]);
@@ -102,10 +113,8 @@ class GapDisnakerController extends Controller
     {
         try {
             $disnaker = GapDisnaker::with('branches')->find($id);
-
-
             $fileName = $request->file('file')->getClientOriginalName();
-            $request->file('file')->storeAs('infra/disnaker/'.$disnaker->id.'/', $fileName, ["disk" => 'public']);
+            $request->file('file')->storeAs('infra/disnaker/' . $disnaker->id . '/', $fileName, ["disk" => 'public']);
 
             $disnaker->file = $fileName;
             $disnaker->save();
